Put new json_dumps tests alongside previous ones
D'oh! Thanks @jbbarth.
